Crate tpm_lib

Travvy’s Project Manager Library (tpm_lib)

tpm_lib serves as the core library for Travvy’s Project Manager (tpm), a command-line utility designed to streamline project management tasks such as opening, creating, listing, editing, and deleting project entries.

This library provides abstractions and utility functions to facilitate user interaction and data manipulation.

Primary Entities

Project: Represents a single project with attributes like name and path.

Action: Enumerates the different actions that can be performed on projects.

DynErr: Represents dynamic errors that can occur within the application.

Dialogue<'a>: Abstraction for handling interactive dialogues.

Features

Interactive Mode: A command-line interface for interactively managing projects.

Project Management: Functions for adding, editing, and deleting project entries.

File Operations: Functions for loading and saving project data from and to disk.

Dependencies

clap- For command-line argument parsing.

serde

serde_json - For serialization and deserialization.

dialoguer - For constructing interactive command-line interfaces.

lazy_static - For lazily-evaluated statics.

Usage

This library is primarily intended to be used by the tpm binary, but it exposes public interfaces that could be utilized in custom extensions or other binaries.

use tpm_lib::{load_projects, Action};

let projects = load_projects();
// Custom logic here

For more examples and usage guidelines, refer to the README.md.

Contribute

For contributing guidelines, please refer to the README.md.

License

This project is licensed under the MIT License - see the LICENSE file for details.

Structs

Project

Enums

Action

Dialogue

DynErr

OpenAction

Constants

ABOUT

APP_NAME

the app name, used everywhere

VALID_SHELLS

VERSION

WELCOME_SCREEN

Statics

HOME_INTERFACE_VISITS

A shared resource that tracks the number of visits to the home interface.

Functions

add_project

change_directory

create_path_with_parent_dirs

delete_project

delete_projects

edit_project

Shows an interface for editing a project and saves the changes.

exit_with

gen_completions

get_config_dir

get_current_shell

gets the current shell from the SHELL environment variable

get_matches

Parses command line arguments and returns a struct containing the parsed values.

get_path_to_shell_profile

get_visits

handler

increment_visits

Increments the number of visits to the home interface by one.

load_projects

load_projects_from_disk

new_project

open_in_editor

open_project

open_projects_file

project_already_exists

quit

save_projects

select_no_projects_found

show_add_project_interface

show_home_interface

show_new_project_interface

show_overwrite_project_interface

show_select_projects_interface