Function tpm_lib::get_matches

pub fn get_matches<I, T>(args: I) -> ArgMatcheswhere
    I: IntoIterator<Item = T>,
    T: Into<OsString> + Clone,

Parses command line arguments and returns a struct containing the parsed values.

Arguments

args - Any value that can be converted into an iterator that yields values that can be converted to OsString. Usually this is std::env::args().

Examples

use clap::ArgMatches;
use tpm_lib::get_matches;

let args = vec!["tpm", "add", "foo", "bar"];
let matches = get_matches(args);

assert_eq!(matches.subcommand_name(), Some("add"));
let add_matches = matches.subcommand_matches("add").unwrap();
assert_eq!(add_matches.value_of("project_name"), Some("foo"));
assert_eq!(add_matches.value_of("project_path"), Some("bar"));
assert_eq!(add_matches.value_of("name"), None);